Formalized the dictionary representation of the bbsa.
{'nodeName(param=value,param2=value2)':[ other nodes]}

This push changed all of the preliminary node storage from lists to dictionaries. This is so I can take a dictionary
representation of the bbsa and reconstruct it easier.

Will be adding loading tomorrow. Had to make a lot of changes. some of them are kind of gross. But it works.

Also added easier capability to select which nodes you want to use in the evolution. This will be useful
once we have a Input GUI.